*|feat: Extract nmr CLI into dedicated `packages/nmr` package
Move all nmr-specific source, bin scripts, and tests from `packages/core` into a new `packages/nmr` package (`@williamthorsen/nmr`). Clean up `packages/core` as an empty shared-library shell with no bin entries, no CLI dependencies, and no `./tests` export.

Rewire all internal references: root `package.json` devDependency, `bootstrap` script, `.config/nmr.config.ts` import, `scripts/postinstall.js` resolution, and release-kit's stale core devDependency. Update `.agents/PROJECT.md` to reflect the new package topology.

Add `passWithNoTests: true` to the shared vitest config so core (now testless) passes the test pipeline.
