+++ config/vitest.config.ts
@@ -15,6 +15,7 @@ export const baseConfig = defineConfig({
       provider: 'v8',
     },
     exclude: ['**/node_modules/**'],
+    passWithNoTests: true, // allow packages with no test files to pass
     watch: false, // don't enter watch mode unless the `--watch` flag is passed
   },
 });
