--- packages/core/src/index.ts
@@ -1,3 +1,0 @@
-export type { NmrConfig } from './config.js';
-export { defineConfig } from './config.js';
-export type { ResolvedContext } from './context.js';
