+++ packages/nmr/__tests__/commands/sync-pnpm-version.test.ts
