+++ packages/nmr/__tests__/config.test.ts
