+++ packages/nmr/__tests__/help.test.ts
