+++ packages/nmr/__tests__/registries.test.ts
