+++ packages/nmr/bin/nmr-sync-pnpm-version.js
