+++ packages/nmr/bin/nmr.js
