+++ packages/nmr/src/cli.ts
