+++ packages/nmr/src/commands/sync-pnpm-version.ts
