+++ packages/nmr/src/help.ts
