+++ packages/nmr/src/helpers/package-json.ts
