+++ packages/nmr/src/helpers/type-guards.ts
