+++ packages/nmr/src/index.ts
@@ -1,0 +1,3 @@
+export type { NmrConfig } from './config.js';
+export { defineConfig } from './config.js';
+export type { ResolvedContext } from './context.js';
