+++ packages/nmr/src/registries.ts
