+++ packages/nmr/src/resolver.ts
