+++ packages/nmr/tsconfig.eslint.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true,
+  },
+  "exclude": [],
+  "include": ["__tests__/", "src/", "eslint.config.js", "*.cjs", "*.json", "*.mjs", "*.mts", "*.ts"],
+}
