+++ packages/nmr/tsconfig.generate-typings.json
@@ -1,0 +1,15 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "noEmit": false,
+    "outDir": "./dist/esm",
+    "rootDir": "./src",
+    "sourceMap": false,
+    "tsBuildInfoFile": "./dist/tsconfig.generate-typings.tsbuildinfo",
+  },
+  "exclude": ["**/__tests__"],
+  "include": ["src"],
+}
