+++ packages/nmr/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "lib": ["es2022"],
+  },
+  "include": ["src/**/*.ts", "src/**/*.d.ts", "*.ts"],
+}
