+++ packages/release-kit/package.json
@@ -44,8 +44,7 @@
     "js-yaml": "4.1.1"
   },
   "devDependencies": {
-    "@types/js-yaml": "4.0.9",
-    "@williamthorsen/node-monorepo-core": "workspace:*"
+    "@types/js-yaml": "4.0.9"
   },
   "engines": {
     "node": ">=18.17.0"
